Week_3_DemoCode/requirements.txt
flask>=3.0.0
flask-smorest>=0.44.0
marshmallow>=3.20.0
pyyaml>=6.0
